news-hub-backend/requirements.txt
fastapi
uvicorn[standard]
motor
httpx
google-generativeai
python-dotenv
pydantic
